+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.app/src/main/java/nts/uk/ctx/at/shared/app/command/remainingnumber/nursingcareleave/UpdateCareLeaveListCommandHandler.java
@@ -30,7 +30,7 @@ public class UpdateCareLeaveListCommandHandler extends CommandHandlerWithResult<
 implements PeregUpdateListCommandHandler<UpdateCareLeaveCommand>{
 	@Inject
 	private ChildCareLeaveRemainingDataService  service;
-	
+
 	@Inject
 	private CareLeaveRemainingInfoRepository careInfoRepo;
 
@@ -39,10 +39,10 @@ implements PeregUpdateListCommandHandler<UpdateCareLeaveCommand>{
 
 	@Inject
 	private ChildCareLeaveRemaiDataRepo childCareDataRepo;
-	
+
 	@Inject
 	private LeaveForCareDataRepo careDataRepo;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00036";
@@ -56,7 +56,7 @@ implements PeregUpdateListCommandHandler<UpdateCareLeaveCommand>{
 	@Override
 	protected List<MyCustomizeException> handle(CommandHandlerContext<List<UpdateCareLeaveCommand>> context) {
 		List<UpdateCareLeaveCommand> cmd = context.getCommand();
-		String cid = AppContexts.user().companyId();		
+		String cid = AppContexts.user().companyId();
 		List<LeaveForCareData> leaveCareDataInsert = new ArrayList<>();
 		List<LeaveForCareData> leaveCareDataUpdate = new ArrayList<>();
 		List<CareLeaveRemainingInfo> leaveCareInfoInsert = new ArrayList<>();
@@ -65,16 +65,16 @@ implements PeregUpdateListCommandHandler<UpdateCareLeaveCommand>{
 		List<ChildCareLeaveRemainingData> childCareDataUpdate = new ArrayList<>();
 		List<ChildCareLeaveRemainingInfo> childCareLeaveInfoInsert = new ArrayList<>();
 		List<ChildCareLeaveRemainingInfo> childCareLeaveInfoUpdate = new ArrayList<>();
-		
+
 		createData(cmd, childCareDataInsert, childCareDataUpdate, leaveCareDataInsert, leaveCareDataUpdate,
 				childCareLeaveInfoInsert, childCareLeaveInfoUpdate, leaveCareInfoInsert, leaveCareInfoUpdate);
-		
+
 		service.addData(cid, childCareDataInsert, leaveCareDataInsert, childCareLeaveInfoInsert, leaveCareInfoInsert);
-		
+
 		service.updateData(cid, childCareDataUpdate, leaveCareDataUpdate, childCareLeaveInfoUpdate, leaveCareInfoUpdate);
 		 return new ArrayList<MyCustomizeException>();
 	}
-	
+
 	private void createData(List<UpdateCareLeaveCommand> cmd,
 			List<ChildCareLeaveRemainingData> childCareDataInsert, List<ChildCareLeaveRemainingData> childCareDataUpdate,
 			List<LeaveForCareData> leaveCareDataInsert, List<LeaveForCareData> leaveCareDataUpdate,
@@ -116,8 +116,8 @@ implements PeregUpdateListCommandHandler<UpdateCareLeaveCommand>{
 					c.getChildCareUseArt().intValue(),
 					c.getChildCareUpLimSet() == null ? UpperLimitSetting.FAMILY_INFO.value
 							: c.getChildCareUpLimSet().intValue(),
-					c.getChildCareThisFiscal() == null ? null : c.getChildCareThisFiscal().doubleValue(),
-					c.getChildCareNextFiscal() == null ? null : c.getChildCareNextFiscal().doubleValue());
+					c.getChildCareThisFiscal() == null ? null : c.getChildCareThisFiscal().intValue(),
+					c.getChildCareNextFiscal() == null ? null : c.getChildCareNextFiscal().intValue());
 			Optional<ChildCareLeaveRemainingInfo> childCareInfoOpt = checkChildCareInfoLst.stream().filter(item -> item.getSId().equals(c.getSId())).findFirst();
 			if (childCareInfoOpt.isPresent()) {
 				childCareLeaveInfoUpdate.add(childCareInfo);
@@ -130,15 +130,15 @@ implements PeregUpdateListCommandHandler<UpdateCareLeaveCommand>{
 					c.getCareUseArt().intValue(),
 					c.getCareUpLimSet() == null ? UpperLimitSetting.FAMILY_INFO.value
 							: c.getCareUpLimSet().intValue(),
-					c.getCareThisFiscal() == null ? null : c.getCareThisFiscal().doubleValue(),
-					c.getCareNextFiscal() == null ? null : c.getCareNextFiscal().doubleValue());
+					c.getCareThisFiscal() == null ? null : c.getCareThisFiscal().intValue(),
+					c.getCareNextFiscal() == null ? null : c.getCareNextFiscal().intValue());
 			Optional<CareLeaveRemainingInfo> careInfoOpt = checkCareInfoLst.stream().filter(item -> item.getSId().equals(c.getSId())).findFirst();
 			if (careInfoOpt.isPresent()) {
 				leaveCareInfoUpdate.add(careInfo);
 			} else {
 				leaveCareInfoInsert.add(careInfo);
 			}
-			
+
 		});
 	}
